% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmlh-package.r
\docType{package}
\name{fmlh-package}
\alias{fmlh-package}
\title{fmlh}
\description{
'fml' is the Fused Matrix Library, a multi-source, header-only C++ library
for dense matrix computing. The emphasis is on real-valued matrix types
for numerical operations useful for data analysis.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
